check queue defined and remove cyclic dependcy
